requirements.txt
fastapi
uvicorn
pymongo
python-jose[cryptography]
python-dotenv